refactor: add retry mechanism for MongoDB operations and handle DuplicateKeyError
